Detection of Various Eye Diseases with Machine Learning Methods Using Deep Features on MATLAB
Nowadays, the detection of eye diseases is carried out by examining them in detail by health professionals specialized in their field. Although rare, it is possible to encounter detection errors in some cases. In the last 4 years in Turkey, the number of cases filed against ophthalmologists as a result of erroneous detection has increased to over 6000. These problems that have occurred are causing Turkey’s trust in the field of health to decrease. To prevent problems like this and this, and experts can help to support that decision for a site, a common 3-eye disease (glaucoma, diabetic retinopathy, Cataracts) with the healthy eye (normal eye) classes will be undertaken. The retinal images sent to the expert in the field are aimed to be as accurate as possible with the prepared deep learning and machine learning algorithms and to support the decision-making mechanism of the expert in the field. The data that will be used in the study “kaggle.com ” it is envisaged to be taken over and studied with a total of about 4500 retinal images. Deep InceptionV3 to work in the first place the removal of features, VGG-16, VGG-19, the use of deep learning models as SqueezeNet, then, support vector machine (SVM),k-nearest neighbor algorithm (k-NN), artificial neural networks (Ann), logistic regression (LR) models, such as classification with machine learning will be studied. As a result of this classification, it is thought that both to support the decision-making mechanisms of doctors who are specialists in their field and to help those who are responsible in the field of health who are not specialists in their field to make more efficient decisions. With the October data to be obtained in the next process, the number of images in the data set can be increased and thus the predicted accuracy rates can be increased. It is planned to analyze the classification results and other graphs to be determined in the study and to make relevant contacts and collaborations for tests in the field of health. As a result of relevant contacts and cooperation, it is aimed to conduct studies on its active use. It is considered that the project has an important place to strengthen the prestige of our country in the field of health.